requirements.txt
py-clob-client>=0.34.0
python-dotenv>=1.0.0
requests>=2.31.0
web3>=6.11.0
eth-account>=0.10.0
colorama>=0.4.6
